+++ b4n-tasks/commands/mod.rs
@@ -10,7 +10,9 @@ pub use self::inject_container::{EphemeralContainerConfig, InjectContainerComman
 pub use self::list_contexts::ListKubeContextsCommand;
 pub use self::list_resource_ports::ListResourcePortsCommand;
 pub use self::list_themes::ListThemesCommand;
-pub use self::new_kubernetes_client::{KubernetesClientError, KubernetesClientResult, NewKubernetesClientCommand};
+pub use self::new_kubernetes_client::{
+    KubernetesClientError, KubernetesClientInfo, KubernetesClientResult, NewKubernetesClientCommand,
+};
 pub use self::run_plugin::{RunPluginCommand, RunPluginError, RunPluginOutput};
 pub use self::save_configuration::SaveConfigurationCommand;
 pub use self::save_content::SaveContentCommand;