+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## WIP
 
-### Bug fixes
+### Features
 
-- fix mouse menu not working properly in YAML edit mode
+- add `--cluster` and `--user` options to override context configuration
+- add `--as` and `--as-group` options to impersonate a user during the session
+- add `--client-cert`, `--client-key`, and `--ca` options for custom TLS authentication
+
+### Bug Fixes
+
+- fix mouse menu not working in YAML edit mode
 
 ## 0.5.6 - 2026-08-17